optimized with lombok

+++ src/main/java/io/github/samirsilva/libraryapi/service/AutorService.java
@@ -1,8 +1,11 @@
 package io.github.samirsilva.libraryapi.service;
 
+import io.github.samirsilva.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.samirsilva.libraryapi.model.Autor;
 import io.github.samirsilva.libraryapi.repository.AutorRepository;
+import io.github.samirsilva.libraryapi.repository.LivroRepository;
 import io.github.samirsilva.libraryapi.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,15 +13,12 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService {
 
     private final AutorRepository repository;
     private final AutorValidator validator;
-
-    public AutorService(AutorRepository repository, AutorValidator validator) {
-        this.repository = repository;
-        this.validator = validator;
-    }
+    private final LivroRepository livroRepository;
 
     public Autor salvar(Autor autor){
         validator.validar(autor);
@@ -38,6 +38,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor){
+        if (possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir um Autor que possui Livro");
+        }
         repository.delete(autor);
     }
 
@@ -52,5 +55,8 @@ public class AutorService {
             return repository.findByNacionalidade(nacionalidade);
         }
         return repository.findAll();
+    }
+    public boolean possuiLivro(Autor autor){
+    return  livroRepository.existsByAutor(autor);
     }
 }
